Version 2 : Add database

+++ order-service/src/main/java/com/training/k8s/model/OrderCreatedEvent.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,6 +16,6 @@ public class OrderCreatedEvent {
     private String orderId;
     private String userId;
     private String productId;
-    private double amount;
+    private BigDecimal amount;
     private LocalDateTime orderDate;
 }
